requirements.txt
flask
tensorflow
numpy
pillow
matplotlib
scikit-learn
